Start + 2 Job view control
