+++ test/Employee/User_Control/Job_Detail.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace test.Employer.User_Control
+{
+    public partial class Job_Detail : UserControl
+    {
+        public Job_Detail()
+        {
+            InitializeComponent();
+        }
+    }
+}
