+++ test/Employer/User_Control/Job_Detail_Edit.cs
@@ -1,0 +1,30 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace test.Employee.User_Control
+{
+    public partial class Job_Detail_Edit : UserControl
+    {
+        public Job_Detail_Edit()
+        {
+            InitializeComponent();
+        }
+
+        private void smallTextBox2_TextChanged(object sender, EventArgs e)
+        {
+
+        }
+
+        private void Job_Detail_Load(object sender, EventArgs e)
+        {
+
+        }
+    }
+}
