+++ test/test.csproj
@@ -83,18 +83,30 @@
     <Compile Include="Employee\Form_Employee_Info.Designer.cs">
       <DependentUpon>Form_Employee_Info.cs</DependentUpon>
     </Compile>
-    <Compile Include="Start\Login.cs">
+    <Compile Include="Employer\User_Control\Job_Detail_Edit.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="Employer\User_Control\Job_Detail_Edit.Designer.cs">
+      <DependentUpon>Job_Detail_Edit.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Employee\User_Control\Job_Detail.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="Employee\User_Control\Job_Detail.Designer.cs">
+      <DependentUpon>Job_Detail.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Start\Start.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Start\Login.Designer.cs">
-      <DependentUpon>Login.cs</DependentUpon>
+    <Compile Include="Start\Start.Designer.cs">
+      <DependentUpon>Start.cs</DependentUpon>
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="User_Control\Available_Job.cs">
+    <Compile Include="Employee\User_Control\Available_Job.cs">
       <SubType>UserControl</SubType>
     </Compile>
-    <Compile Include="User_Control\Available_Job.Designer.cs">
+    <Compile Include="Employee\User_Control\Available_Job.Designer.cs">
       <DependentUpon>Available_Job.cs</DependentUpon>
     </Compile>
     <EmbeddedResource Include="Employee\Form_Job_Employee.resx">
@@ -112,8 +124,14 @@
     <EmbeddedResource Include="Employee\Form_Employee_Info.resx">
       <DependentUpon>Form_Employee_Info.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Start\Login.resx">
-      <DependentUpon>Login.cs</DependentUpon>
+    <EmbeddedResource Include="Employer\User_Control\Job_Detail_Edit.resx">
+      <DependentUpon>Job_Detail_Edit.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Employee\User_Control\Job_Detail.resx">
+      <DependentUpon>Job_Detail.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Start\Start.resx">
+      <DependentUpon>Start.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
@@ -125,7 +143,7 @@
       <DependentUpon>Resources.resx</DependentUpon>
       <DesignTime>True</DesignTime>
     </Compile>
-    <EmbeddedResource Include="User_Control\Available_Job.resx">
+    <EmbeddedResource Include="Employee\User_Control\Available_Job.resx">
       <DependentUpon>Available_Job.cs</DependentUpon>
     </EmbeddedResource>
     <None Include="packages.config" />
@@ -142,9 +160,7 @@
   <ItemGroup>
     <None Include="App.config" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Employer\" />
-  </ItemGroup>
+  <ItemGroup />
   <ItemGroup>
     <None Include="Resources\background.jpg" />
   </ItemGroup>
